--- src/restore-step.ts
@@ -1,7 +1,0 @@
-import restore from "./restore";
-
-async function runRestoreStep(): Promise<void> {
-    await restore();
-}
-
-runRestoreStep();
